requirements.txt
discord.py[voice]>=1.3.4
html2text>=2020.1.16
python-dotenv>=0.14.0
youtube_dl>=2020.7.28
